+++ libraries/wuss/test/tasks/ball.c
@@ -56,6 +56,7 @@ result_t ball_create(wuss_t *wuss, ball_task_t *task)
   rc = wuss_task_create(wuss, &delegate_desc, &delegate);
   if (rc != result_OK)
     return rc;
+  wuss_task_set_autoclose(delegate, 1);
 
   return wuss_window_create_placed(delegate,
                                    SIZE2D(200, 160),
@@ -219,7 +220,7 @@ result_t ball_handle(wuss_window_t      *window,
   case wuss_EVENT_IDLE:
     return ball_idle(task_data);
 
-  case wuss_EVENT_CLOSE:
+  case wuss_EVENT_QUIT:
     free(bc); /* task_data was calloc'd per instance by the spawner */
     return result_OK;