+++ libraries/wuss/test/tasks/sofa.c
@@ -376,6 +376,7 @@ result_t sofa_create(wuss_t*wuss, sofa_task_t*task)
   rc = wuss_task_create(wuss, &delegate_desc, &delegate);
   if (rc != result_OK)
     return rc;
+  wuss_task_set_autoclose(delegate, 1);
 
   return wuss_window_create_placed(delegate,
                                    SIZE2D(180, 160),
@@ -588,7 +589,7 @@ result_t sofa_handle(wuss_window_t      *window,
   case wuss_EVENT_IDLE:
     return sofa_idle(task_data);
 
-  case wuss_EVENT_CLOSE:
+  case wuss_EVENT_QUIT:
     free(sc); /* task_data was calloc'd per instance by the spawner */
     return result_OK;